backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.0.0
pydub==0.25.1
python-dotenv==1.0.0
lxml==5.1.0
pillow==11.3.0
gunicorn==21.2.0
